refactor(duty-tracker): move error types to separate module
