+++ crates/duty-tracker/src/lib.rs
@@ -5,6 +5,7 @@
 pub mod contract_manager;
 pub mod contract_persister;
 pub mod contract_state_machine;
+pub mod errors;
 pub mod predicates;
 pub mod stake_chain_persister;
 pub mod stake_chain_state_machine;
